pyproject.toml
[project]
name = "roms-regrid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Luke", email = "gregorl@ethz.ch" }
]
requires-python = ">=3.12"
dependencies = [
    "dask>=2024.4.1",
    "loguru>=0.7.3",
    "munch>=4.0.0",
    "numpy>=1.26.4",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.1",
    "xarray>=2024.3.0",
    "xesmf>=0.8.5",
]

[project.scripts]
roms-regrid = "roms_regrid:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
